acceptInvitation is now done through swal. Deleted obsolete svelte file.
